new google map template project
I created a new Google map template project which had a search bar and
map. The search bar uses a places autocomplete  module that drops down
a list of suggested places while time. When a user presses enter or
clicks on a suggested item, it loads the place on the map after
fetching the place details data.  *work in progress*